Floor variances after reloading model for mllr. Thanks to Xiaochuan Niu.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@10132 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ sphinx3/include/s3_cfg.h
@@ -109,7 +109,7 @@ extern "C" {
 
 #define s3_cfg_id2index(x) (x & S3_CFG_INDEX_MASK)
 
-typedef uint32 s3_cfg_id_t;
+typedef size_t s3_cfg_id_t;
 
 typedef struct s3_cfg_rule_s {
   s3_cfg_id_t src;
